+++ kipper/core/src/compiler/analysis/symbol-table/function-scope.ts
@@ -38,7 +38,7 @@ export class FunctionScope extends LocalScope {
 	public addArgument(declaration: ParameterDeclaration): ScopeParameterDeclaration {
 		const identifier = declaration.getSemanticData().identifier;
 
-		// Ensuring that the declaration does not overwrite other definitions
+		// Ensuring that the declaration does not overwrite other declarations
 		this.ctx.programCtx.semanticCheck(declaration).identifierNotUsed(identifier, this);
 
 		const scopeDeclaration = new ScopeParameterDeclaration(declaration);