+++ kipper/core/src/compiler/ast/nodes/expressions/arithmetic/index.ts
@@ -1,0 +1,7 @@
+/**
+ * Arithmetic expression module containing the classes representing expressions, which can be used to perform
+ * calculations based on two expressions.
+ * @since 0.11.0
+ */
+export * from "./additive-expression";
+export * from "./multiplicative-expression";