+++ kipper/core/src/compiler/ast/nodes/index.ts
@@ -1,9 +1,3 @@
-/**
- * Language-specific AST nodes enriched with semantic data and logical handling that implement the semantic
- * analysis, type checking and translation of Kipper code.
- * @since 0.8.0
- */
+export * from "./declarations";
 export * from "./expressions";
-export * from "./definitions";
 export * from "./statements";
-export * from "./factories";