+++ kipper/core/src/compiler/ast/nodes/statements/index.ts
@@ -1,0 +1,13 @@
+/**
+ * Statement module containing the classes representing statements, which are used to perform specific operations and
+ * define the flow of a program.
+ * @since 0.11.0
+ */
+export * from "./statement";
+export * from "./iteration/";
+export * from "./compound-statement";
+export * from "./expression-statement";
+export * from "./if-statement";
+export * from "./jump-statement";
+export * from "./return-statement";
+export * from "./switch-statement";