+++ kipper/core/src/compiler/ast/semantic-data/index.ts
@@ -1,5 +1,5 @@
 /**
- * Semantic data definitions for all AST nodes.
+ * Semantic data declarations for all AST nodes.
  * @since 0.10.0
  */
 export * from "./definitions";